requirements.txt
streamlit==1.44.1
diffusers==0.32.2
torch==2.6.0
pandas==2.2.3
numpy==2.0.2
pillow==11.1.0
gdown==5.2.0